Filter demographic data by selected filtering parameters

Filter demographic data by selected filtering parameters

Usage

filter_demo(
  data,
  data_type = c("t_variant", "t_invariant"),
  filter_param,
  id_col = NULL,
  any = FALSE,
  rm_na = TRUE,
  log_path = NULL
)

Arguments

data

A data frame containing pre-processed demographic data.

data_type

A character string. Type of demographic data: "t_variant" or "t_invariant"

filter_param

A named list containing filtering parameters. The names in the list are the column names and the values are vectors of values to keep.

id_col

A character string. Name of ID column in data set.

Optional, necessary only when any = TRUE

any

Logical. Filtering option, any year. Default is FALSE

rm_na

Logical. Should rows with NA in the non-filtered columns be removed? Default is FALSE

If TRUE, removes observations that have NA in any of the non-filtered columns.

log_path

A character string. Path to the log file to append function logs. Default is NULL

If NULL, a new directory /log and file is created in the current working directory.

Value

Filtered demographic dataframe containing only relevant observations based on the filtering parameters.

Examples

# Filter varying and unvarying datasets

log_file <- tempfile()
cat("Example log file", file = log_file)

filtered_var <- filter_demo(data = var_df,
data_type = "t_variant",
filter_param = list("year_varying" = c(2012:2015), "varying_code" = c("1146")),
log_path = log_file)
#> Filtering time-variant dataset...
#> ✔ Filtered time-variant by 'year_varying and varying_code' column(s)
#> ℹ Filtered 269648 rows (99.8% removed)
#> 
#> ! The dataset has no NAs or they are coded in a different format.
#> 
#> ────────────────────────────────────────────────────────────────────────────────
#> Demographic dataset succesfully filtered
#> 
#> 
#> ── Data Summary ────────────────────────────────────────────────────────────────
#> 
#> ── After filtering: 
#> ℹ Remaining number of rows: 568
#> ℹ Remaining number of columns: 3
#> 
#> tibble [568 × 3] (S3: tbl_df/tbl/data.frame)
#>  $ id          : chr [1:568] "P000000037" "P000000037" "P000000037" "P000002133" ...
#>  $ year_varying: int [1:568] 2012 2013 2014 2012 2013 2014 2015 2012 2013 2014 ...
#>  $ varying_code: chr [1:568] "1146" "1146" "1146" "1146" ...

filtered_invar <- filter_demo(data = invar_df, data_type = "t_invariant",
filter_param = list("y_birth" = c(2006:2008),
"innvandringsgrunn" = c("FAMM", "UTD")),
rm_na = FALSE,
log_path = log_file)
#> Filtering time-invariant dataset...
#> ✔ Filtered time-invariant dataset by 'y_birth and innvandringsgrunn' column(s)
#> ℹ Filtered 21169 rows (70.5% removed)
#> 
#> ────────────────────────────────────────────────────────────────────────────────
#> Demographic dataset succesfully filtered
#> 
#> 
#> ── Data Summary ────────────────────────────────────────────────────────────────
#> 
#> ── After filtering: 
#> ℹ Remaining number of rows: 8855
#> ℹ Remaining number of columns: 4
#> 
#> tibble [8,855 × 4] (S3: tbl_df/tbl/data.frame)
#>  $ id               : chr [1:8855] "P000000111" "P000000183" "P000000431" "P000000882" ...
#>  $ sex              : Factor w/ 2 levels "0","1": 2 1 2 2 1 1 1 2 2 2 ...
#>  $ y_birth          : int [1:8855] 2006 2008 2006 2008 2008 2007 2007 2007 2006 2007 ...
#>  $ innvandringsgrunn: chr [1:8855] "FAMM" "UTD" "FAMM" "FAMM" ...