Read and validate the structure of demographic individual-level data

read_demo_data() validates the general structure and minimum column requirements for demographic individual-level data. The input data sets must be CSV, RDS, RDA or .SAV files.

Usage

read_demo_data(
  file_path,
  data_type = c("t_variant", "t_invariant"),
  id_col = "id",
  date_col = "date",
  log_path = NULL,
  ...
)

Arguments

file_path

A character string. File path to the demographic data to read. Supports CSV, RDS, RDA and .SAV files.

data_type

A character string. Demographic data can either be of type "t_variant" or "t_invariant", necessary to check correct data structure characteristics.

id_col

A character string. Name of ID column in data set. Default is "id".

date_col

A character string. Name of date column in data set, default is "date".

log_path

A character string. Path to the log file to append function logs. Default is NULL.

If NULL, a new directory /log and file is created in the current working directory.

...

Additional arguments passed to methods or underlying functions.

Value

A data frame with the validated minimum requirements for demographic data.

Examples

# Read and validate CSV file for varying individual level demographic data
demo_csv <- system.file("extdata", "invar_data.csv", package = "regtools")
log_file <- tempfile()
cat("Example log file", file = log_file)

demo_data_validated <- read_demo_data(demo_csv, data_type = "t_invariant",
id_col = "id", log_path = log_file)
#> 
#> Reading C:/Users/ALMS/AppData/Local/Temp/RtmpAN06X8/temp_libpath155cf797f81/regtools/extdata/invar_data.csv file...
#> ✔ Succesfully read file: C:/Users/ALMS/AppData/Local/Temp/RtmpAN06X8/temp_libpath155cf797f81/regtools/extdata/invar_data.csv
#> Checking column requirements:
#> ✔ ID column
#> ✔ Date column
#> 
#> Data type: time invariant. Checking requirements...
#> ✔ No duplicate IDs
#> 
#> 
#> ────────────────────────────────────────────────────────────────────────────────
#> Demographic dataset succesfully read and columns validated
#> 
#> 
#> ── Data Summary ────────────────────────────────────────────────────────────────
#> 
#> ℹ Number of rows: 30024. Number of columns: 4.
#> ℹ Unique IDs in dataset: 30024.
#> 
#> 'data.frame':	30024 obs. of  4 variables:
#>  $ id               : chr  "P000000037" "P000000052" "P000000059" "P000000111" ...
#>  $ sex              : int  0 0 0 1 1 0 1 0 0 0 ...
#>  $ y_birth          : int  2008 2000 2007 2003 2000 2003 2009 2005 2004 2002 ...
#>  $ innvandringsgrunn: chr  "FAMM" "FAMM" "FAMM" "FAMM" ...